+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- add validation for npm executable on `bit doctor`
 - ignore `import`/`require` statements from CDN (http/https)
 - avoid generating package.json inside node_modules for author when one of the component files is package.json
 - preserve indentation of `package.json` files and default to 2 spaces, similar to NPM (#1630)